:hammer: assume sources fields are markdown instead of HTML
