+++ packages/@ourworldindata/utils/src/index.ts
@@ -608,6 +608,7 @@ export {
     getUrlTarget,
     checkIsInternalLink,
     convertHeadingTextToId,
+    markdownToEnrichedTextBlock,
 } from "./GdocsUtils.js"
 
 export {
